backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
ffmpeg-python==0.2.0
Pillow==10.3.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
watchdog==4.0.0
aiofiles==23.2.1
python-multipart==0.0.9
httpx==0.27.0
python-telegram-bot==21.5
